Most basic 3d rendering with z buffer is implemented and working

+++ GrEn/Window.h
@@ -28,7 +28,7 @@ struct windowEvent
 
 struct frameExtra
 {
-	int z;
+	float z;
 };
 
 class Camera;
